requirements.txt
py-cord==2.4.1
aiohttp==3.9.1
aiosqlite==0.19.0
psutil==5.9.6
python-dotenv==1.0.0